requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
greenlet==2.0.1
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.44
tornado==6.1
tzdata==2022.7
tzlocal==4.2
